refactor directory and files structure in app-domain component
fix respective imports in app-reports and in app-analyses
