+++ app-analyses/src/main/kotlin/com/usvision/analyses/analyzer/Analyzer.kt
@@ -1,7 +1,7 @@
 package com.usvision.analyses.analyzer
 
-import com.usvision.model.Visitable
-import com.usvision.model.Visitor
+import com.usvision.model.visitor.Visitable
+import com.usvision.model.visitor.Visitor
 
 interface Analyzer<out T> : Visitor {
     fun getResults(): Map<Visitable, T>
