+++ app-analyses/src/main/kotlin/com/usvision/analyses/analyzer/MicroserviceBelongingAnalyzer.kt
@@ -1,9 +1,9 @@
 package com.usvision.analyses.analyzer
 
-import com.usvision.model.CompanySystem
-import com.usvision.model.Database
-import com.usvision.model.Microservice
-import com.usvision.model.Visitable
+import com.usvision.model.domain.CompanySystem
+import com.usvision.model.domain.databases.Database
+import com.usvision.model.domain.Microservice
+import com.usvision.model.visitor.Visitable
 
 class MicroserviceBelongingAnalyzer : RelationshipAnalyzer {
     private val parentOf: MutableMap<Visitable, Relationship> = mutableMapOf()
