+++ app-analyses/src/test/kotlin/com/usvision/analyses/analyzer/MicroserviceBelongingAnalyzerTest.kt
@@ -1,7 +1,7 @@
 package com.usvision.analyses.analyzer
 
-import com.usvision.model.CompanySystem
-import com.usvision.model.Microservice
+import com.usvision.model.domain.CompanySystem
+import com.usvision.model.domain.Microservice
 import kotlin.test.*
 
 internal class MicroserviceBelongingAnalyzerTest {
