--- app-model/src/main/kotlin/com/usvision/model/Visitor.kt
@@ -1,11 +1,0 @@
-package com.usvision.model
-
-interface Visitable {
-    fun accept(visitor: Visitor)
-}
-
-interface Visitor {
-    fun visit(companySystem: CompanySystem)
-    fun visit(microservice: Microservice)
-    fun visit(database: Database)
-}
