+++ app-model/src/main/kotlin/com/usvision/model/domain/databases/Database.kt
@@ -1,0 +1,8 @@
+package com.usvision.model.domain.databases
+
+import com.usvision.model.visitor.Visitable
+
+interface Database : Visitable {
+    val id: String?
+    val description: String
+}
