+++ app-model/src/main/kotlin/com/usvision/model/domain/operations/Operation.kt
@@ -1,0 +1,5 @@
+package com.usvision.model.domain.operations
+
+abstract class Operation(
+    open val description: String = ""
+)
