+++ app-model/src/main/kotlin/com/usvision/model/domain/operations/RestEndpoint.kt
@@ -1,0 +1,7 @@
+package com.usvision.model.domain.operations
+
+data class RestEndpoint(
+    val httpVerb: String,
+    val path: String,
+    override val description: String = ""
+) : Operation(description)
