+++ app-model/src/main/kotlin/com/usvision/model/visitor/Visitable.kt
@@ -1,0 +1,5 @@
+package com.usvision.model.visitor
+
+interface Visitable {
+    fun accept(visitor: Visitor)
+}
