+++ app-reports/src/main/kotlin/com/usvision/reports/SystemRepository.kt
@@ -1,6 +1,6 @@
 package com.usvision.reports
 
-import com.usvision.model.System
+import com.usvision.model.systemcomposite.System
 
 interface SystemRepository {
     fun load(name: String): System
